implement basis for staff crud UI
The vector of staff members is shared data between all tabs and is
passed into the view/update function of the management tab, which also
has private data to hold the state of UI elements like inputs/buttons
for each staff member in a vector.
These vectors have to be combined to implement one row for each staff
member to update their cardid etc. Matching up the lifetimes of
everything was the hardest part, key insight was that if 'a is the
lifetime of the management tab's private data and 'b the lifetime of the
shared data, then 'a > 'b. At first I set them equal which lead to
problems in the view function of main.rs. (although maybe we can
actually do that if we rewrite the view function to explicitly borrow
the management/timetrack fields of the Stechuhr structure)

In the end I also had to remove the `type Message` definition of the Tab
trait because I had an error E0309 which apparently wanted me to add a
lifetime to the message but that did not make any sense to me. I don't
understand the purpose of the type definition anyways, at first I
thought it's supposed to break some sort of circularity but the program
also compiles without it.