implement user crud and use custom types in db
I finally figured out what diesel wants from me.
Basically if you have a struct A and enum B and want to derive
AsChangeset, Insertable, Queryable for A, you need to

- set `#[sql_type = "Sometype"]` for B, which declares the type by which the enum will be
represented in the db
- derive AsExpression for B, which just uses sql_type to implement a
trait that AsChangeset then uses
- implement FromSql/ToSql for B
- maybe implement queryable for A by hand, not sure yet.

ToSql/FromSql deal with individual values and Queryable seems to be a
way to load an entire struct from a database row (which then implies
that the struct must correspond exactly to a row, e.g. no joins but
maybe there is a nother trait for that)
I implemented Queryable for WorkEventT by hand because I want to ignore
the primary key.